MFC 196478:
Prior to the dire warning about values of network_interfaces other than
AUTO the biggest mistake users made was leaving lo0 off the list. Since
lo0 is effectively mandatory, check for it and add it to the list if
it's not there.

MFC 196523:
Improve the case test to detect the presence of lo0 in the list of
network_interfaces.
Submitted by:	Christoph Mallon <christoph.mallon@gmx.de>

Approved by:	re (kib)


git-svn-id: http://svn.freebsd.org/base/stable/8@196645 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f