MFC	r201544
An existing incomplete ARP entry would expire a subsequent
statically configured entry of the same host. This bug was
due to the expiration timer was not cancelled when installing
the static entry. Since there exist a potential race condition
with respect to timer cancellation, simply check for the
LLE_STATIC bit inside the expiration function instead of
cancelling the active timer.


git-svn-id: http://svn.freebsd.org/base/stable/8@202132 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f