This is an MFC of 205698
Clean up some printing stuff so that we can have a bit finer control
on debug output. Add a new platform function requirement to allow
for printing based upon the ITL nexus instead of the isp unit plus
channel, target and lun. This allows some printouts and error messages
from the core code to appear in the same format as the platform's
subsystem (in FreeBSD's case, CAM path).


git-svn-id: http://svn.freebsd.org/base/stable/8@206196 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f