MFC r207553, r207555, r207651:
Implement the no_user_check option to pam_krb5.

This option is available in the Linux implementation of pam_krb5
and allows to authorize a user not known to the local system.

Ccache is not used as we don't have a secure uid/gid for the cache file.

Usable for authentication of external kerberos users (e.g Active Directory)
via PAM from applications like Cyrus saslauthd, PHP or perl.

PR:		bin/146186
Approved by:	deplhij (mentor)


git-svn-id: http://svn.freebsd.org/base/stable/8@208193 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f