Belatedly MFC r210002, r210676, r210812, r210826:
r210002:
In the example for how to create a VLAN, also include an example of
setting the IP address.  While it is documented earlier in rc.conf(5)
that the '.' in the VLAN name becomes a '_' in rc.conf, this may not be
easy to find when just using rc.conf(5) as reference documentation.

r210676:
Fix a bunch of typos and spelling mistakes.

r210812:
Update references from nonexistent usbconfig(1) to usbconfig(8).

r210826:
Correctly sort usbconfig(8) within the SEE ALSO section.

PR:                         docs/149221
Submitted by:               Lars Hartmann (lars at chaotika dot org)
Help with mergeinfo from:   kib
Reviewed by:                kib, hrs (IPv6 parts that should not be
merged to stable/8)


git-svn-id: http://svn.freebsd.org/base/stable/8@212828 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f