Partially MFC r215166:
Disable priming the congestion window from the host cache. The current method
interacts poorly with delayed ack and appropriate byte counting amongst other
things, resulting in undesired delay during a connection's opening slow start.
Even if we did fix the issues, the current method is still dubious at best and
needs to be thought through thoroughly.

Due to a mistake on my behalf, the change described above was committed to head
as part of a larger patch in revision 215166. Instead of waiting for the MFC of
215166, I'm merging just this small portion for the upcoming release without
bringing the mergeinfo for r215166 along. The mergeinfo will sort itself out
when r215166 is eventually merged.

This is an intentional direct commit to the 8-STABLE branch.

Reported by:	Maxim Dounin <mdounin at mdounin ru> and others
Submitted by:	andre


git-svn-id: http://svn.freebsd.org/base/stable/8@215925 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f